+++ src/Akeneo/Platform/Bundle/UIBundle/Resources/public/js/product/form/quantified-associations/components/QuantifiedAssociationRow.tsx
@@ -95,6 +95,7 @@ const RowAction = styled.div`
 type QuantifiedAssociationRowProps = {
   row: Row;
   parentQuantifiedLink: QuantifiedLink | undefined;
+  isUserOwner?: boolean;
   isCompact?: boolean;
   onChange: (row: Row) => void;
   onRemove: (row: Row) => void;
@@ -104,6 +105,7 @@ const QuantifiedAssociationRow = ({
   row,
   parentQuantifiedLink,
   isCompact = false,
+  isUserOwner = true,
   onChange,
   onRemove,
 }: QuantifiedAssociationRowProps) => {
@@ -113,8 +115,10 @@ const QuantifiedAssociationRow = ({
   const productEditUrl = useRoute(`pim_enrich_${row.productType}_edit`, {id: row.product?.id.toString() || ''});
   const thumbnailUrl = useProductThumbnail(row.product);
   const blueColor = useTheme().color.blue100;
-  const canRemoveAssociation = isGranted('pim_enrich_associations_remove') && undefined === parentQuantifiedLink;
-  const canUpdateQuantity = isGranted('pim_enrich_associations_edit') && isGranted('pim_enrich_associations_remove');
+  const canRemoveAssociation =
+    isGranted('pim_enrich_associations_remove') && undefined === parentQuantifiedLink && isUserOwner;
+  const canUpdateQuantity =
+    isGranted('pim_enrich_associations_edit') && isGranted('pim_enrich_associations_remove') && isUserOwner;
 
   const handleQuantityChange = (event: ChangeEvent<HTMLInputElement>) => {
     if (!canUpdateQuantity) return;